set IValidator<D> for Validation -> for Commands Handler

+++ SmsHub.Application/Features/Config/Handlers/Commands/Delete/Implementations/CcSendDeleteHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using SmsHub.Application.Features.Config.Handlers.Commands.Delete.Contracts;
 using SmsHub.Common.Extensions;
 using SmsHub.Domain.Features.Config.MediatorDtos.Commands.Delete;
@@ -12,10 +13,12 @@ namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementa
         private readonly IMapper _mapper;
         private readonly ICcSendCommandService _ccSendCommandService;
         private readonly ICcSendQueryService _ccSendQueryService;
+        private readonly IValidator<DeleteCcSendDto> _validator;
         public CcSendDeleteHandler(
             IMapper mapper,
             ICcSendCommandService ccSendCommandService,
-            ICcSendQueryService ccSendQueryService)
+            ICcSendQueryService ccSendQueryService,
+            IValidator<DeleteCcSendDto> validator)
         {
             _mapper=mapper;
             _mapper.NotNull(nameof(mapper));
@@ -25,9 +28,18 @@ namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementa
 
             _ccSendQueryService=ccSendQueryService;
             _ccSendQueryService.NotNull(nameof(ccSendQueryService));
+
+            _validator=validator;   
+            _validator.NotNull(nameof(validator));
         }
         public async Task Handle(DeleteCcSendDto deleteCcSendDto, CancellationToken cancellationToken)
         {
+            var validationResult=await _validator.ValidateAsync(deleteCcSendDto, cancellationToken);
+            if (!validationResult.IsValid)
+            {
+                throw new InvalidDataException();
+            }
+
             var ccSend=await _ccSendQueryService.Get(deleteCcSendDto.Id);
             _ccSendCommandService.Delete(ccSend);
         }
