+++ SmsHub.Application/Features/Config/Handlers/Commands/Delete/Implementations/ConfigTypeGroupDeleteHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using SmsHub.Application.Features.Config.Handlers.Commands.Delete.Contracts;
 using SmsHub.Common.Extensions;
 using SmsHub.Domain.Features.Config.MediatorDtos.Commands.Delete;
@@ -12,10 +13,12 @@ namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementa
         private readonly IMapper _mapper;
         private readonly IConfigTypeGroupCommandService _configTypeGroupCommandService;
         private readonly IConfigTypeGroupQueryService _configTypeGroupQueryService;
+        private readonly IValidator<DeleteConfigTypeGroupDto> _validator;
         public ConfigTypeGroupDeleteHandler(
             IMapper mapper,
             IConfigTypeGroupCommandService configTypeGroupCommandService,
-            IConfigTypeGroupQueryService configTypeGroupQueryService)
+            IConfigTypeGroupQueryService configTypeGroupQueryService,
+            IValidator<DeleteConfigTypeGroupDto> validator)
         {
             _mapper = mapper;
             _mapper.NotNull(nameof(mapper));
@@ -25,9 +28,18 @@ namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementa
 
             _configTypeGroupQueryService = configTypeGroupQueryService;
             _configTypeGroupQueryService.NotNull(nameof(configTypeGroupQueryService));
+
+            _validator = validator;
+            _validator.NotNull(nameof(validator));
         }
         public async Task Handle(DeleteConfigTypeGroupDto deleteConfigTypeGroupDto, CancellationToken cancellationToken)
         {
+            var validationResult=await _validator.ValidateAsync(deleteConfigTypeGroupDto, cancellationToken);
+            if (!validationResult.IsValid)
+            {
+                throw new InvalidDataException();
+            }
+
             var configTypeGroup = await _configTypeGroupQueryService.Get(deleteConfigTypeGroupDto.Id);
             _configTypeGroupCommandService.Delete(configTypeGroup);
         }
