+++ SmsHub.Application/Features/Config/Handlers/Commands/Delete/Implementations/DisallowedPhraseDeleteHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using SmsHub.Application.Features.Config.Handlers.Commands.Delete.Contracts;
 using SmsHub.Common.Extensions;
 using SmsHub.Domain.Features.Config.MediatorDtos.Commands.Delete;
@@ -7,13 +8,17 @@ using SmsHub.Persistence.Features.Config.Queries.Contracts;
 
 namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementations
 {
-    public class DisallowedPhraseDeleteHandler: IDisallowedPhraseDeleteHandler
+    public class DisallowedPhraseDeleteHandler : IDisallowedPhraseDeleteHandler
     {
         private readonly IMapper _mapper;
         private readonly IDisallowedPhraseCommandService _disallowedPhraseCommandService;
         private readonly IDisallowedPhraseQueryService _disallowedPhraseQueryService;
+        private readonly IValidator<DeleteDisallowedPhraseDto> _validator;
         public DisallowedPhraseDeleteHandler(
-            IMapper mapper, IDisallowedPhraseCommandService disallowedPhraseCommandService, IDisallowedPhraseQueryService disallowedPhraseQueryService)
+            IMapper mapper,
+            IDisallowedPhraseCommandService disallowedPhraseCommandService,
+            IDisallowedPhraseQueryService disallowedPhraseQueryService,
+            IValidator<DeleteDisallowedPhraseDto> validator)
         {
             _mapper = mapper;
             _mapper.NotNull(nameof(mapper));
@@ -23,10 +28,19 @@ namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementa
 
             _disallowedPhraseQueryService = disallowedPhraseQueryService;
             _disallowedPhraseQueryService.NotNull(nameof(disallowedPhraseQueryService));
+
+            _validator = validator;
+            _validator.NotNull(nameof(validator));
         }
         public async Task Handle(DeleteDisallowedPhraseDto deleteDisallowedPhraseDto, CancellationToken cancellationToken)
         {
-            var disallowedPhrase=await _disallowedPhraseQueryService.Get(deleteDisallowedPhraseDto.Id);
+            var validationResult = await _validator.ValidateAsync(deleteDisallowedPhraseDto, cancellationToken);
+            if (!validationResult.IsValid)
+            {
+                throw new InvalidDataException();
+            }
+
+            var disallowedPhrase = await _disallowedPhraseQueryService.Get(deleteDisallowedPhraseDto.Id);
             _disallowedPhraseCommandService.Delete(disallowedPhrase);
         }
     }
