+++ SmsHub.Application/Features/Config/Handlers/Commands/Delete/Implementations/PermittedTimeDeleteHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using SmsHub.Application.Features.Config.Handlers.Commands.Delete.Contracts;
 using SmsHub.Common.Extensions;
 using SmsHub.Domain.Features.Config.MediatorDtos.Commands.Delete;
@@ -7,15 +8,17 @@ using SmsHub.Persistence.Features.Config.Queries.Contracts;
 
 namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementations
 {
-    public class PermittedTimeDeleteHandler: IPermittedTimeDeleteHandler
+    public class PermittedTimeDeleteHandler : IPermittedTimeDeleteHandler
     {
         private readonly IMapper _mapper;
         private readonly IPermittedTimeCommandService _permittedTimeCommandService;
         private readonly IPermittedTimeQueryService _permittedTimeQueryService;
+        private readonly IValidator<DeletePermittedTimeDto> _validator;
         public PermittedTimeDeleteHandler(
-            IMapper mapper, 
-            IPermittedTimeCommandService permittedTimeCommandService, 
-            IPermittedTimeQueryService permittedTimeQueryService)
+            IMapper mapper,
+            IPermittedTimeCommandService permittedTimeCommandService,
+            IPermittedTimeQueryService permittedTimeQueryService,
+            IValidator<DeletePermittedTimeDto> validator)
         {
             _mapper = mapper;
             _mapper.NotNull(nameof(mapper));
@@ -25,10 +28,19 @@ namespace SmsHub.Application.Features.Config.Handlers.Commands.Delete.Implementa
 
             _permittedTimeQueryService = permittedTimeQueryService;
             _permittedTimeQueryService.NotNull(nameof(permittedTimeQueryService));
+
+            _validator = validator;
+            _validator.NotNull(nameof(validator));
         }
         public async Task Handle(DeletePermittedTimeDto deletePermittedTimeDto, CancellationToken cancellationToken)
         {
-            var permittedTime=await _permittedTimeQueryService.Get(deletePermittedTimeDto.Id);
+            var validationResult = await _validator.ValidateAsync(deletePermittedTimeDto, cancellationToken);
+            if (!validationResult.IsValid)
+            {
+                throw new InvalidDataException();
+            }
+
+            var permittedTime = await _permittedTimeQueryService.Get(deletePermittedTimeDto.Id);
             _permittedTimeCommandService.Delete(permittedTime);
         }
     }
