+++ SmsHub.Application/Features/Contact/Handlers/Commands/Delete/Implementations/ContactCategoryDeleteHandler.cs
@@ -1,21 +1,28 @@
 ﻿using AutoMapper;
+using Azure.Core;
+using FluentValidation;
 using SmsHub.Application.Features.Contact.Handlers.Commands.Delete.Contracts;
 using SmsHub.Common.Extensions;
+using SmsHub.Domain.Features.Contact.MediatorDtos.Commands.Create;
 using SmsHub.Domain.Features.Contact.MediatorDtos.Commands.Delete;
 using SmsHub.Persistence.Features.Contact.Commands.Contracts;
 using SmsHub.Persistence.Features.Contact.Queries.Contracts;
+using System.Threading;
 
 namespace SmsHub.Application.Features.Contact.Handlers.Commands.Delete.Implementations
 {
-    public class ContactCategoryDeleteHandler: IContactCategoryDeleteHandler
+    public class ContactCategoryDeleteHandler : IContactCategoryDeleteHandler
     {
         private readonly IMapper _mapper;
         private readonly IContactCategoryCommandService _contactCategoryCommandService;
         private readonly IContactCategoryQueryService _contactCategoryQueryService;
-        public ContactCategoryDeleteHandler( 
+        private readonly IValidator<DeleteContactCategoryDto> _validator;
+
+        public ContactCategoryDeleteHandler(
             IMapper mapper,
             IContactCategoryCommandService contactCategoryCommandService,
-            IContactCategoryQueryService contactCategoryQueryService)
+            IContactCategoryQueryService contactCategoryQueryService,
+            IValidator<DeleteContactCategoryDto> validator)
         {
             _mapper = mapper;
             _mapper.NotNull(nameof(mapper));
@@ -25,9 +32,18 @@ namespace SmsHub.Application.Features.Contact.Handlers.Commands.Delete.Implement
 
             _contactCategoryQueryService = contactCategoryQueryService;
             _contactCategoryQueryService.NotNull(nameof(contactCategoryQueryService));
+
+            _validator = validator;
+            _validator.NotNull(nameof(_validator));
         }
         public async Task Handle(DeleteContactCategoryDto deleteContactCategoryDto, CancellationToken cancellationToken)
         {
+            var validationResult = await _validator.ValidateAsync(deleteContactCategoryDto, cancellationToken);
+            if (!validationResult.IsValid)
+            {
+                throw new InvalidDataException();
+            }
+
             var contactCategory = await _contactCategoryQueryService.Get(deleteContactCategoryDto.Id);
             _contactCategoryCommandService.Delete(contactCategory);
         }
