+++ SmsHub.Application/Features/Contact/Handlers/Commands/Delete/Implementations/ContactNumberDeleteHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using SmsHub.Application.Features.Contact.Handlers.Commands.Delete.Contracts;
 using SmsHub.Common.Extensions;
 using SmsHub.Domain.Features.Contact.MediatorDtos.Commands.Delete;
@@ -12,10 +13,12 @@ namespace SmsHub.Application.Features.Contact.Handlers.Commands.Delete.Implement
         private readonly IMapper _mapper;
         private readonly IContactNumberCommandService _contactNumberCommandService;
         private readonly IContactNumberQueryService _contactNumberQueryService;
+        private readonly IValidator<DeleteContactNumberDto> _validator;
         public ContactNumberDeleteHandler(
             IMapper mapper, 
             IContactNumberCommandService contactNumberCommandService, 
-            IContactNumberQueryService contactNumberQueryService)
+            IContactNumberQueryService contactNumberQueryService,
+            IValidator<DeleteContactNumberDto> validator)
         {
             _mapper = mapper;
             _mapper.NotNull(nameof(mapper));
@@ -25,9 +28,18 @@ namespace SmsHub.Application.Features.Contact.Handlers.Commands.Delete.Implement
 
             _contactNumberQueryService = contactNumberQueryService;
             _contactNumberQueryService.NotNull(nameof(contactNumberQueryService));
+
+            _validator = validator;
+            _validator.NotNull(nameof(validator));
         }
         public async Task Handle(DeleteContactNumberDto deleteContactNumberDto, CancellationToken cancellationToken)
         {
+            var validationResult=await _validator.ValidateAsync(deleteContactNumberDto, cancellationToken);
+            if (!validationResult.IsValid)
+            {
+                throw new InvalidDataException();
+            }
+
             var contactNumber = await _contactNumberQueryService.Get(deleteContactNumberDto.Id);
             _contactNumberCommandService.Delete(contactNumber);
         }
